refactor: restructure composables and remove unused global registry
