+++ src/runtime/composables/index.ts
@@ -1,4 +1,4 @@
-import { useRoute, useRequestHeaders, useCookie } from '#imports'
+import { useRoute, useRequestHeaders, useCookie, useRouter } from '#imports'
 import { ref, computed, watch, onUnmounted } from 'vue'
 import { parseAcceptLanguage } from '../internal'
 import { nuxtI18nInternalOptions, nuxtI18nOptionsDefault, localeCodes as _localeCodes } from '#build/i18n.options.mjs'
@@ -6,24 +6,34 @@ import { getActiveHead } from 'unhead'
 import { useI18n } from 'vue-i18n'
 
 import type { Ref } from 'vue'
-import type { DetectBrowserLanguageOptions, SeoAttributesOptions } from '#build/i18n.options.mjs'
+import type {
+  DetectBrowserLanguageOptions,
+  I18nHeadMetaInfo,
+  I18nHeadOptions,
+  SeoAttributesOptions
+} from '#build/i18n.options.mjs'
 
 export * from 'vue-i18n'
 export * from './shared'
 
-import type { Locale } from 'vue-i18n'
 import { getNormalizedLocales, type HeadParam } from '../utils'
 import {
   getAlternateOgLocales,
   getCanonicalLink,
   getCurrentOgLocale,
   getHreflangLinks,
-  getOgUrl
+  getOgUrl,
+  getRouteBaseName,
+  localeHead,
+  localeLocation,
+  localePath,
+  localeRoute,
+  switchLocalePath
 } from '../routing/compatibles'
 import { findBrowserLocale, getLocale, getLocales } from '../routing/utils'
-import { useLocaleHead as _useLocaleHead } from '../routing/composables'
 
-export { useRouteBaseName, useLocalePath, useLocaleRoute, useSwitchLocalePath } from '../routing/composables'
+import type { Locale } from 'vue-i18n'
+import type { RouteLocation, RouteLocationNormalizedLoaded, RouteLocationRaw, Router } from 'vue-router'
 
 /**
  * Returns a function to set i18n params.
@@ -37,7 +47,6 @@ export { useRouteBaseName, useLocalePath, useLocaleRoute, useSwitchLocalePath } 
 export type SetI18nParamsFunction = (params: Record<string, unknown>) => void
 export function useSetI18nParams(seoAttributes?: SeoAttributesOptions): SetI18nParamsFunction {
   const route = useRoute()
-  // const router = useRouter()
   const head = getActiveHead()
 
   const i18n = useI18n()
@@ -105,24 +114,208 @@ export function useSetI18nParams(seoAttributes?: SeoAttributesOptions): SetI18nP
 /**
  * The `useLocaleHead` composable returns localized head properties for locale-related aspects.
  *
- * @param options - An options object, see {@link I18nHeadOptions}.
+ * @param options - An options, see about details {@link I18nHeadOptions}
  *
  * @returns The localized {@link I18nHeadMetaInfo | head properties} with Vue `ref`.
  *
  * @public
  */
-export function useLocaleHead(
-  options?: Pick<
-    NonNullable<Parameters<typeof _useLocaleHead>[0]>,
-    'addDirAttribute' | 'addSeoAttributes' | 'identifierAttribute'
-  >
-): ReturnType<typeof _useLocaleHead> {
-  const { addDirAttribute, addSeoAttributes, identifierAttribute } = options || {}
-  return _useLocaleHead({
-    addDirAttribute: addDirAttribute || false,
-    addSeoAttributes: addSeoAttributes || false,
-    identifierAttribute: identifierAttribute || 'hid'
+export function useLocaleHead({
+  addDirAttribute = false,
+  addSeoAttributes = false,
+  identifierAttribute = 'hid'
+}: I18nHeadOptions = {}): Ref<I18nHeadMetaInfo> {
+  const router = useRouter()
+
+  const metaObject: Ref<I18nHeadMetaInfo> = ref({
+    htmlAttrs: {},
+    link: [],
+    meta: []
   })
+
+  function cleanMeta() {
+    metaObject.value = {
+      htmlAttrs: {},
+      link: [],
+      meta: []
+    }
+  }
+
+  function updateMeta() {
+    metaObject.value = localeHead({ addDirAttribute, addSeoAttributes, identifierAttribute }) as I18nHeadMetaInfo
+  }
+
+  if (process.client) {
+    const stop = watch(
+      () => router.currentRoute.value,
+      () => {
+        cleanMeta()
+        updateMeta()
+      },
+      { immediate: true }
+    )
+    onUnmounted(() => stop())
+  } else {
+    updateMeta()
+  }
+
+  return metaObject
+}
+
+/**
+ * The function that resolves the route base name.
+ *
+ * @remarks
+ * The parameter signatures of this function is the same as {@link getRouteBaseName}.
+ *
+ * @param givenRoute - A route location. The path or name of the route or an object for more complex routes.
+ *
+ * @returns The route base name, if route name is not defined, return `null`.
+ *
+ * @see {@link useRouteBaseName}
+ *
+ * @public
+ */
+export type RouteBaseNameFunction = (givenRoute?: RouteLocationNormalizedLoaded) => string | undefined
+
+/**
+ * The `useRouteBaseName` composable returns a function which returns the route base name.
+ *
+ * @remarks
+ * The function returned by `useRouteBaseName` is the wrapper function with the same signature as {@link getRouteBaseName}.
+ *
+ * @returns A {@link RouteBaseNameFunction}.
+ *
+ * @public
+ */
+export function useRouteBaseName(): RouteBaseNameFunction {
+  return getRouteBaseName
+}
+
+/**
+ * The function that resolve locale path.
+ *
+ * @remarks
+ * The parameter sygnatures of this function is same as {@link localePath}.
+ *
+ * @param route - A route location. The path or name of the route or an object for more complex routes.
+ * @param locale - A locale optional, if not specified, uses the current locale.
+ *
+ * @returns Returns the localized URL for a given route.
+ *
+ * @see {@link useLocalePath}
+ *
+ * @public
+ */
+export type LocalePathFunction = (route: RouteLocation | RouteLocationRaw, locale?: Locale) => string
+
+/**
+ * The `useLocalePath` composable returns function  that resolve the locale path.
+ *
+ * @remarks
+ * The function returned by `useLocalePath` is the wrapper function with the same signature as {@link localePath}.
+ *
+ * @returns A {@link LocalePathFunction}.
+ *
+ * @public
+ */
+export function useLocalePath(): LocalePathFunction {
+  return localePath
+}
+
+/**
+ * The function that resolve route.
+ *
+ * @remarks
+ * The parameter sygnatures of this function is same as {@link localeRoute}.
+ *
+ * @param route - A route location. The path or name of the route or an object for more complex routes.
+ * @param locale - A locale optinal, if not specified, uses the current locale.
+ *
+ * @returns the route object for a given route, the route object is resolved by vue-router rather than just a full route path.
+ *
+ * @see {@link useLocaleRoute}
+ *
+ * @public
+ */
+export type LocaleRouteFunction = (
+  route: RouteLocationRaw,
+  locale?: Locale
+) => ReturnType<Router['resolve']> | undefined
+
+/**
+ * The `useLocaleRoute` composable returns function that resolve the locale route.
+ *
+ * @remarks
+ * The function returned by `useLocaleRoute` is the wrapper function with the same signature as {@link localeRoute}.
+ *
+ * @returns A {@link LocaleRouteFunction}.
+ *
+ * @public
+ */
+export function useLocaleRoute(): LocaleRouteFunction {
+  return localeRoute
+}
+
+/**
+ * The function that resolve locale location.
+ *
+ * @remarks
+ * The parameter sygnatures of this function is same as {@link localeLocation}.
+ *
+ * @param route - A route location. The path or name of the route or an object for more complex routes.
+ * @param locale - A locale optional, if not specified, uses the current locale.
+ *
+ * @returns the location object for a given route, the location object is resolved by vue-router rather than just a full route path.
+ *
+ * @see {@link useLocaleLocation}
+ *
+ * @public
+ */
+export type LocaleLocationFunction = (route: RouteLocationRaw, locale?: Locale) => Location | RouteLocation | undefined
+
+/**
+ * The `useLocaleLocation` composable returns function that resolve the locale location.
+ *
+ * @remarks
+ * The function returned by `useLocaleLocation` is the wrapper function with the same signature as {@link localeLocation}.
+ *
+ * @returns A {@link LocaleLocationFunction}.
+ *
+ * @public
+ */
+export function useLocaleLocation(): LocaleLocationFunction {
+  return localeLocation
+}
+
+/**
+ * The functin that swtich locale path.
+ *
+ * @remarks
+ * The parameter sygnatures of this function is same as {@link switchLocalePath}.
+ *
+ * @param locale - A locale optional, if not specified, uses the current locale.
+ *
+ * @returns A link to the current route in another language.
+ *
+ * @see {@link useSwitchLocalePath}
+ *
+ * @public
+ */
+export type SwitchLocalePathFunction = (locale: Locale) => string
+
+/**
+ * The `useSwitchLocalePath` composable returns function that resolve the locale location.
+ *
+ * @remarks
+ * The function returned by `useSwitchLocalePath` is the wrapper function with the same signature as {@link switchLocalePath}.
+ *
+ * @returns A {@link SwitchLocalePathFunction}.
+ *
+ * @public
+ */
+export function useSwitchLocalePath(): SwitchLocalePathFunction {
+  return switchLocalePath
 }
 
 /**
