--- src/runtime/routing/composables/index.ts
@@ -1,2 +1,0 @@
-export * from './head'
-export * from './routing'
